Convert explicit default bindings removal to a feature flag
The feature flag is called `implicit_default_bindings`. It has the same
role as the previously existing upgrade function. The upgrade function
is removed in the process.

Compared to the upgrade function however, it does remove entries in the
`rabbit_route`, `rabbit_reverse_route` and `rabbit_semi_durable_route`
tables in addition to `rabbit_durable_route` table. Also, the list of
queues is taken from the `rabbit_queue` table instead of the
`rabbit_durable_queue` one.

[#163224049,#159298729]